+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * [UIBULKED-459](https://folio-org.atlassian.net/browse/UIBULKED-459) Bulk Edit Actions for Holdings Notes - Staff Only for Added Notes.
 * [UIBULKED-479](https://folio-org.atlassian.net/browse/UIBULKED-479) Extend User search in User lookup.
 * [UIBULKED-445](https://folio-org.atlassian.net/browse/UIBULKED-445) Populate Bulk Edit Form with Instance Note Types.
+* [UIBULKED-446](https://folio-org.atlassian.net/browse/UIBULKED-446) Bulk Edit Actions for All Instance Notes - Add and Remove Notes.
 
 ## [4.1.0](https://github.com/folio-org/ui-bulk-edit/tree/v4.1.0) (2024-03-19)