+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## In progress
 
+* [UIBULKED-493](https://folio-org.atlassian.net/browse/UIBULKED-493) Rename "Find (full field search)" to "Find" for bulk edit Email in Users record type.
+
 ## [4.1.3](https://github.com/folio-org/ui-bulk-edit/tree/v4.1.3) (2024-04-19)
 
 * [UIBULKED-478](https://folio-org.atlassian.net/browse/UIBULKED-478) Update permissions related to export manager module.